refactor: Move all methods out from GrandpaService and into GrandpaRound.
