+++ src/main/java/com/limechain/network/protocol/grandpa/GrandpaMessageHandler.java
@@ -82,11 +82,6 @@ public class GrandpaMessageHandler {
         }
 
         GrandpaRound round = roundCache.getRound(voteMessageSetId, voteMessageRoundNumber);
-        if (round == null) {
-            round = new GrandpaRound();
-            round.setRoundNumber(voteMessageRoundNumber);
-            roundCache.addRound(voteMessageSetId, round);
-        }
 
         SubRound subround = signedMessage.getStage();
         switch (subround) {
@@ -238,7 +233,7 @@ public class GrandpaMessageHandler {
 
             // Check if needed to catch-up peer
             if (neighbourMessage.getRound().compareTo(
-                    grandpaSetState.fetchLatestRound().getRoundNumber().add(CATCH_UP_THRESHOLD)) >= 0) {
+                    grandpaSetState.fetchLatestRoundNumber().add(CATCH_UP_THRESHOLD)) >= 0) {
                 log.log(Level.FINE, "Neighbor message indicates that the round of Peer " + peerId + " is ahead.");
 
                 CatchUpReqMessage catchUpReqMessage = CatchUpReqMessage.builder()
@@ -270,7 +265,7 @@ public class GrandpaMessageHandler {
             throw new GrandpaGenericException("Catch up message has a different setId.");
         }
 
-        if (catchUpReqMessage.getRound().compareTo(grandpaSetState.fetchLatestRound().getRoundNumber()) > 0) {
+        if (catchUpReqMessage.getRound().compareTo(grandpaSetState.fetchLatestRoundNumber()) > 0) {
             throw new GrandpaGenericException("Catching up on a round in the future.");
         }
